Generators: Embedd into correct event when using collision-context
This fixes a problem with embedding and when option `--embedIntoFile`
is used. Here, we had an index problem in the primary generator class, which
was not taking into account the actual event index according the collision context.
This could lead to wrong vertices or using wrong MCHeader information for the embedding decision.